Bit of cleanup
