--- app/mvp/page.module.css
@@ -1,154 +1,0 @@
-.main {
-  height: 100vh;
-}
-
-.nodeContainer {
-  position: absolute;
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  padding: 1rem;
-  width: var(--size-width-node);
-  height: var(--size-height-node);
-  border-style: solid;
-  border-radius: var(--size-border-radius-node);
-  border-width: var(--size-border-width-node);
-  border-color: var(--color-border-node);
-  cursor: grab;
-  user-select: none;
-}
-
-.numberNodeInput {
-  text-align: center;
-  font-size: var(--size-font);
-  font-weight: bold;
-  height: var(--size-height-input);
-  border-style: solid;
-  border-radius: var(--size-border-radius-node);
-  border-width: var(--size-border-width-node);
-  border-color: var(--color-border-node);
-  background-color: var(--color-background-input);
-}
-
-.numberNodeSlider {
-  height: 4px;
-  background-color: var(--color-border-node);
-  border-width: var(--size-border-width-node);
-  border-radius: var(--size-border-radius-node);
-  appearance: none;
-}
-
-.numberNodeSlider::-webkit-slider-thumb {
-  -webkit-appearance: none;
-  appearance: none;
-  padding: 0;
-  width: var(--size-width-port);
-  height: var(--size-height-port);
-  border-style: solid;
-  border-radius: var(--size-border-radius-round);
-  border-width: var(--size-border-width-node);
-  border-color: var(--color-border-node);
-  background-color: var(--color-background-numberNode);
-}
-
-.outputPort {
-  position: absolute;
-  top: 100%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  height: var(--size-height-port);
-  width: var(--size-width-port);
-  border-style: solid;
-  border-radius: var(--size-border-radius-round);
-  border-width: var(--size-border-width-node);
-  border-color: var(--color-border-node);
-  background-color: var(--color-background-numberNode);
-}
-
-.outputPort:hover {
-  cursor: crosshair;
-}
-
-.inputPort1 {
-  position: absolute;
-  top: 0%;
-  left: 25%;
-  padding: 0;
-  transform: translate(-50%, -50%);
-  height: var(--size-height-port);
-  width: var(--size-width-port);
-  border-style: solid;
-  border-radius: var(--size-border-radius-round);
-  border-width: var(--size-border-width-node);
-  border-color: var(--color-border-node);
-  background-color: var(--color-background-mathNode);
-}
-
-.inputPort1:hover {
-  cursor: crosshair;
-}
-
-.inputPort2 {
-  position: absolute;
-  top: 0%;
-  left: 75%;
-  padding: 0;
-  transform: translate(-50%, -50%);
-  height: var(--size-height-port);
-  width: var(--size-width-port);
-  border-style: solid;
-  border-radius: var(--size-border-radius-round);
-  border-width: var(--size-border-width-node);
-  border-color: var(--color-border-node);
-  background-color: var(--color-background-mathNode);
-}
-
-.inputPort2:hover {
-  cursor: crosshair;
-}
-
-.mathNodeLabel {
-  display: flex;
-  justify-content: center;
-  font-weight: bold;
-  font-size: var(--size-font);
-}
-
-.mathNodeValue {
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  height: var(--size-height-input);
-  font-size: var(--size-font);
-  font-weight: bold;
-  border-style: solid;
-  border-radius: var(--size-border-radius-node);
-  border-width: var(--size-border-width-node);
-  border-color: var(--color-border-node);
-  background-color: var(--color-background-input);
-  cursor: text;
-  user-select: text;
-}
-
-.connectors {
-  position: absolute;
-  height: 100%;
-  width: 100%;
-  pointer-events: none;
-}
-
-.nodeContainer[data-node-type="number"] {
-  background-color: var(--color-background-numberNode);
-}
-
-.nodeContainer[data-node-type="math"] {
-  background-color: var(--color-background-mathNode);
-}
-
-#stop-1 {
-  stop-color: var(--color-background-numberNode);
-}
-
-#stop-2 {
-  stop-color: var(--color-background-mathNode);
-}
