--- components/canvas/styles.module.css
@@ -1,3 +1,0 @@
-.main {
-  height: 100vh;
-}
