--- components/canvas/types.ts
@@ -1,6 +1,0 @@
-export interface CurrentPath {
-  x1?: number;
-  y1?: number;
-  x2?: number;
-  y2?: number;
-}
