--- components/svg/types.ts
@@ -1,3 +1,0 @@
-import type { SVGProps } from "react";
-
-export type SvgProps = SVGProps<SVGElement>;
